refactor: remove client parser and stick with default parser
The client parser being removed and the server parser being
made the de facto was because of the differences between the
parsers as well as the parsed output.

The server parser (now default) is built on top of `css.parse()`:
https://github.com/reworkcss/css/blob/master/lib/parse/index.js

The client parser took advantage of the browser style property:
https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/style

Several differences were encountered:
- hexadecimal colors were converted to rgb in the browser
- quotes were stripped from declaration values in the browser
- prefixing behavior was inconsistent between environments

Thus, sticking with the regex parser is to ensure isomorphism
going forward.